refactor(CarNameValidatorTest, TryNumberValidatorTest): 검증에 대한 테스트 코드 패키지로 분리
